Adding user based validation to vaultService

+++ backend/src/main/java/com/mealchemy/vault/service/VaultService.java
@@ -3,22 +3,34 @@ package com.mealchemy.vault.service;
 /* Import libraries */
 import org.springframework.stereotype.Service;
 import java.util.stream.Collectors;
-import java.util.List;
+import java.util.*;
+import org.springframework.web.server.*;
+import org.springframework.http.*;
 
 /* Import classes */
 import com.mealchemy.vault.model.Vault;
+import com.mealchemy.auth.model.User;
 import com.mealchemy.vault.dto.VaultResponse;
 import com.mealchemy.vault.dto.VaultRequest;
 import com.mealchemy.vault.repository.VaultRepository;
+import com.mealchemy.vault.repository.VaultMemberRepository;
+import com.mealchemy.auth.repository.UserRepository;
+import com.mealchemy.shared.enums.VaultType;
 
 @Service
 public class VaultService
 {   
     private final VaultRepository vaultRepository;
 
-    public VaultService(VaultRepository vaultRepository)
+    private final VaultMemberRepository vaultMemberRepository;
+
+    private final UserRepository userRepository;
+
+    public VaultService(VaultRepository vaultRepository, VaultMemberRepository vaultMemberRepository, UserRepository userRepository)
     {
         this.vaultRepository = vaultRepository;
+        this.vaultMemberRepository = vaultMemberRepository;
+        this.userRepository = userRepository;
     }
 
     // Get all vaults that belong to ownerId
@@ -28,23 +40,49 @@ public class VaultService
     }
 
     // Get a single vault using id
-    public VaultResponse getVault(int id)
+    public VaultResponse getVault(int id, Integer userId)
     {
-        Vault vaultForReturn = vaultRepository.findById(id).orElseThrow(() -> new RuntimeException("Vault not found."));
+        Vault vaultForReturn = vaultRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Vault not found."));
+
+        boolean isOwner = vaultForReturn.getOwnerId().equals(userId);
+        boolean isMember = vaultMemberRepository.existsByVault_VaultIdAndUser_UserId(id, userId);
+
+        if (!isOwner && !isMember)
+        {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only a vault member/owner can view it.");
+        }
+
         return VaultResponse.from(vaultForReturn);
     }
 
     // Post to create a new vault
-    public VaultResponse createVault(VaultRequest request, Integer ownerId)
+    public VaultResponse createVault(VaultRequest request, Integer userId)
     {
-        Vault vaultForReturn = mapRequestToEntity(request, ownerId);
+        User userToCheck = userRepository.findById(userId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found."));
+        
+        if(request.vaultType().equals(VaultType.PRIVATE))
+        {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Users only get one private vault.");
+        }
+
+        Vault vaultForReturn = mapRequestToEntity(request, userId);
         return VaultResponse.from(vaultRepository.save(vaultForReturn));
     }
 
     // Put to update an existing vault
     public VaultResponse updateVault(int id, VaultRequest request, Integer ownerId)
     {
-        Vault vaultForReturn = vaultRepository.findById(id).orElseThrow(() -> new RuntimeException("Vault not found."));
+        Vault vaultForReturn = vaultRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Vault not found."));
+
+        if (!vaultForReturn.getOwnerId().equals(ownerId))
+        {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Vault can only be edited by the owner.");
+        }
+
+        if(request.vaultType().equals(VaultType.PRIVATE))
+        {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Users only get one private vault.");
+        }
 
         vaultForReturn.setOwnerId(ownerId);
         vaultForReturn.setVaultType(request.vaultType());
@@ -54,8 +92,20 @@ public class VaultService
     }
 
     // Delete a specific vault using id
-    public void deleteVault(int id)
+    public void deleteVault(int id, Integer ownerId)
     {
+        Vault vaultToCheck = vaultRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Vault not found."));
+
+        if (!vaultToCheck.getOwnerId().equals(ownerId))
+        {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Vaults can only be deleted be the owner.");
+        }
+
+        if (vaultToCheck.getVaultType().equals(VaultType.PRIVATE))
+        {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Private vaults can't be deleted.");
+        }
+
         vaultRepository.deleteById(id);
     }
